Fixed concurrency issue
git-svn-id: http://equinox:10080/_svn_root/AObjectServer@21 1a28b8b1-d4e8-a449-a3f8-e3d44feecac6

+++ AObjectServer/AOS_Base/AOSContextQueueThreadPool.cpp
@@ -6,7 +6,7 @@
 
 AOSContextQueueThreadPool::AOSContextQueueThreadPool(
   AOSServices& services,
-  int threadCount,                    // = 1
+  size_t threadCount,                 // = 1
   AOSContextQueueInterface *pForward, // = NULL, 
   AOSContextQueueInterface *pBack,    // = NULL
   AOSContextQueueInterface *pError    // = NULL
